moved library code back to root from sub-folder
when github was adding a hash to the library download zip, there was confusion from beginners about needing to rename the unzipped folder before moving into the arduino ide.
this library has been prepped (thijse) for the arduino library manager, which requires the library to be in the root directory.  Having the library in the manager also removes the "beginner user hash" concern.